refactor: Use fetch-result to retrieve and save report
part of story #38263 Choose my own columns based on field name
(numeric,text, dates)

How to test:
No functional change expected in cross-tracker search widgets.

Why?
fetch-result lets us be more flexible and expressive for fetch
operations: we can clearly express that it can fail for various reasons.
We are forced to deal with errors, and we can use Fault to represent
erors instead of string error messages.

This does not replace completely all fetch operations as it would be too
big. The rest of the fetch operations will be done later.

Change-Id: I83d016aa89b0603b1215ed03ea5edb749f29549f