feat: add getResponse and suffix fetch result methods with "Response"
Implements request #38300 Add "Response" suffix to fetch-result methods

How to test:
- No functional change in the Timetracking personal widget. In the
"details" modal, it still retrieves the list of times entered by the
user and shows a "Load more" button. You can simulate that by entering a
bunch of times and changing the "limit" parameter to 2.
- No functional change expected elsewhere.

Why?
- getResponse is useful in cases like the timetracking widget because we
can do just 1 call to the API to get the total number of items and
retrieve the first batch of items.
- Naming is hard, I chose to suffix the methods to suggest their return
type: either JSON is decoded, or a Response is returned. Note that some
methods that return a Response such as del(), head() and options() are
NOT renamed because it's not possible for them to return anything else
than a Response. There is never any body for them to decode.

Change-Id: Id52e1180d0b501aa290ebdc6172b83bd61705c1d