feat: Introduce dropdown for section
This dropdown will allow many actions on the current section
* get the anchored link to the section so that I can share it with my
  BFF
* go to the corresponding artifact
* edit the section
* remove the section

For now only "go to artifact" and "edit" are available.

Go to artifact: we decide to not display the artifact id (ids are too
internal), and not display the tracker name (we are already seeing a
"bug", we want to go to the artifact of the bug, not the bug) (this
removes the management of plural of the tracker name: "Go to bugs").

Edit: the option is not available if user do not have the rights. The
option is disabled if the section is already being edited.

To avoid z-index nightmares, they are now centered in a common file, so
that we don't have to search for every definition of z-index in the app.

The dropdown can be accessed via keyboard ([Tab] [Tab] … then [Space]).

Part of story #37543: edit Artidoc sections

Change-Id: Ib1c100be959f69010b0dc39e2ee6cccf39930a77