feat: Allow configuration of artidoc
User can use REST API to configure the tracker of a document.
This will be used later to create a new section.

PUT artidoc/:id/configuration
{
  selected_tracker_ids: [<tracker-id>]
}

One and only one tracker id is accepted for now:
* We don't know yet if we want to allow to select "No tracker" for a
  document
* In the future we might have more than one type of section (I could
  create a document with both Story and Bugs tracker)
To anticipate the breaking change, we accept an array of tracker ids,
but limited to 1 element.

If the given tracker is not found, deleted, or not viewable by current
user, then a 403 is raised.

The document item must exist, and user must have write access.

The selected tracker is given to the app mount point as data attribute.
You can view source to see the result.

part of story #38288: add a new section

Change-Id: I48c97e8c96e6924fb2878b5762b9d1e636914fa2