refactor: Use fetch-result to get projects
part of story #38263 Choose my own columns based on field name
(numeric,text, dates)

How to test:
No functional change in cross-tracker search widgets. When you click on
the widget to choose trackers, it will load the list of projects sorted
by project label. It will load it only once for all widgets (as before).

Why?
fetch-result lets us be more flexible and expressive for fetch
operations: we can clearly express that it can fail for various reasons.
We are forced to deal with errors, and we can use Fault to represent
errors instead of string error messages.

This does not replace completely all fetch operations as it would be too
big. The rest of the fetch operations will be done later.

Change-Id: I4c657766487e2380fa216f9d07e2827f1fe488b7