refactor: Extract the predefined period selection into the lib
Part of story #37941 Timetracking management widget

Extraction of the predefined period selector from the library.
Before this commit, the selector had to be in every component that needed to use it, and this represented a lot of duplicated code.
Now we've extracted the selector ans its events, into the dedicated library and created a custom component with Hybrid.
The predefined period selector is used in `Personal-Timetracking` and `Timetracking-Management` widgets.

No functional changes expected

Change-Id: Ieea0a318a6d28a78a33ae4bb389fbd2b99c3a435