refactor: Replace tlp-fetch by fetch-result
part of story #38263 Choose my own columns based on field name
(numeric,text, dates)

How to test:
- Interpolated error messages should no longer have wrong encoding for
accented characters: when searching a query, when retrieving the list of
projects.
- When you load the list of trackers, and there is an error, it is still
shown in the alert at the top of the widget.
- When you click on the "CSV export" button, it will offer to download a
CSV file with the list of artifacts matching the report and their fields
matched by duck-typing. When there is an error, it is shown in the
alert.

Why?
fetch-result lets us be more flexible and expressive for fetch
operations: we can clearly express that it can fail for various reasons.
We are forced to deal with errors, and we can use Fault to represent
errors instead of string error messages.
Also note that gettext functions in vue3-gettext are able to interpolate
directly. They also HTML-encode the parameters by default, so we have to
disable this for all error handling cases, otherwise accented characters
are wrongly HTML-encoded.

This finishes replacing all fetch operations. tlp-fetch is removed.

Change-Id: I4fb41fff08e40efdfac965b3cf84f5f9f43bf1fb