feat: Display add new section button
This contribution only introduce the new section button. It involves
some refactoring in the display of sections in order to have correct
management of background colors with this new button: since we still
keep the ol/li semantic, the button must be part of the li, but must not
have the background color of the section (e.g.: section in edit mode).
Another alternative would have been to remove the usage of ol/li and use
some div tags soup, but we are not at this stage yet.

By default the button is not displayed. In order to test it, change the
line in `DocumentContent.vue` to enable the display of the button, and
rebuild/watch the application.
```
const has_add_button = true;
```

There should be a button above each section, and after the last one.

Empty documents does not have add section button because it is a special
case : users will be able to directly enter section details without
having to press a button (one click less).

Change-Id: I9ed65a6b72bb08bffe6147b1126d0be3e8b72800