feat: Force configuration of document if empty
When a document is empty, and user can edit the document, then the
configuration screen should be displayed.

Since we should avoid modal whenever possible, the configuration is done
directly in the page so that the user flow is not blocked.

No success is displayed yet since we will introduce directly the new
section form later once implemented. We will see then what should be
done.

No functional change for the modal.

Part of story #38288: add a new section

Change-Id: I7031a9f7c96e4102f79309d385bd1c1731489ce6