feat: configure tracker before adding a new section
When user clicks on the Add new section button, if the document is not
configured with a tracker, then the modal is displayed in order to let
the user complete the configuration.

If the document is configured, then the button does nothing yet, in
future contribution it will display the new section form. (It seems odd
to have a button that does nothing but we are confident that the new
section form will come before the code freeze. If it is not the case
then we will add a flag to deactivate the button.)

Part of story #38288: add a new section

Change-Id: I82936be1f57f88298518f2473f72f28f16f01199