chore: step 1 - hybrids: 8.2.2 -> 9.0.0
part of request #38309 hybrids: 8.2.2 -> 9.0.0

Hybrids has been bumped to 9.0.0 for the following apps and libraries:
- `@tuleap/plugin-cardwall-card-fields` (uses only the `html` function)
- `@tuleap/commonmark-popover` (used in pull-requests comments)
- `@tuleap/lazybox`
- `@tuleap/plugin-pullrequest-homepage` (unit-tests only)
- `@tuleap/plugin-pullrequests-app` (file-diff placeholders)
- `@tuleap/plugin-pullrequest-selectors-dropdown` (used in
  pull-request-homepage)


Note:

`hybrids` is no longer marked as an external dependency of the artifact modal in order to avoid a clash between hybrids@8.2.2 (modal) and hybrids@9.0.0 (lazybox). It will be marked back as an external dependency once the artifact-modal will be migrated to hybrids@9.0.0.

-- no functional change expected

Change-Id: I2680e140a76ca54b41c1db22cee3308d996e745c