refactor: remove Primitive Obsession
Until now we were relying on a selected tracker id. Instead we should
rely on the tracker itself instead of only the id.

This will allow us to have the possibility later to introduce contextual
information about the tracker. For example we would like to know which
fields id to use when we create a new section (those information will be
added in the brand new DocumentTrackerRepresentation).

No functional change in the document configuration.

Part of story #38288: add a new section

Change-Id: Icf60e56136cf0ee123f0ae35fc8e0c5a87e13067