feat: provide title and description information
In order to be able to create new sections, user must have submit
permission on title and description field.

In order to avoid the "taskboard effect", where users don't know why the
plus button is not displayed, here a message is displayed for the user
in the configuration modal. It is only a warning and not an error
because the permissions differ from user to user according to their
memberships.

To test remove the submit permission for title, description, or both.
Open the configuration modal => you should see a warning.

Part of story #38288: add a new section

Change-Id: I4c8495de83f03e95f42522df248f86b5362984af