feat: display new section form
When user clicks on the Add new section button, then a form is
displayed. The save is not implemented yet so it will end up in error.
If user cancel, then the new section is removed.

No changes for the existing sections.

Note: the display of the new section form for empty document will be
done in a dedicated changeset.

Change-Id: I60bad5b5ce853b36af4d867d50263f1199b0b5da