feat: Display new section form for empty document
This contribution is a follow-up to previous contribution[0], in order
to manage empty documents.

No changes when document is not empty.

Given I have write permission on an empty document
When I open the document
And the document is already configured with a tracker
And I can submit title and description
Then the editor is open to enter a new section.

Given I have write permission on an empty document
When I open the document
And the document is already configured with a tracker
And I cannnot submit title or description
Then the empty state is displayed.

Given I have write permission on an empty document
When I open the document
And the document is not configured with a tracker
And I choose a tracker
And save the configuration
Then the editor is open to enter a new section.

Given I have only read permission on an empty document
When I open the document
And the document is empty
Then the empty state is displayed.

Note: took the opportunity to replace InjectionKey by StrictInjectionKey
in artidoc.

[0]: git #tuleap/stable/e017ce74b5469e09184d496c30a8f37d78c69c2e

Change-Id: If4f8ca0437761971c7e1fd4cc5606119afcd7a5c