feat: Be able to select users through Lazybox
Part of story #37941 Timetracking management widget

Added a user selector to the query editor to let you choose the users for whom you want to see times.
Users are not retained on a new search, persistence will come in dedicated commit.
No data is currently retrieved, as this commit only allows users to be selected.

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

How to test:
Click on the dates. You need to type at least 3 characters of a name in the search input to see users in the drop-down list.
Click on search, user avatars are supposed to be displayed in the query viewer.

Change-Id: I300f7d4e0c8e643d5d6ef696f0e3a7c5bb439c56