feat: Add new section in artidoc app
Now when the user clicks on [save] button for a new section, the section
is properly saved. The TOC is updated with the new title, the section is
highlighted as success and the editor is closed.

Note: the management of page leave whith open editor is broken, it will
be managed in a dedicated contribution.

Part of story #38288: add a new section

Change-Id: Ibd73153eb14bcd7454e5c6dfb9a6fa5dd623dcf6