refactor: split useSectionEditor
We put all the editor code in one file, so I split it into smaller contexts.
Now, we can identify the dependencies of each context and their states.
The next steps might involve using a store to manage states and using pure functions.

Part of story #38288: add a new section

Change-Id: I48f4ae124a8059183e6d1e5a365ba4c2c4181423