refactor: introduce SectionIdentifier
There are two issue with actual management of sections' id:
1. The UUID is build by the DAO and exposed. This is an issue because
UUID lives under the `Tuleap\Db` namespace. We should not have a
dependency on `Db` everywhere in the call stack.
2. The DAO deals with two different id. The UUID when an element is
retrieved or inserted in the database. The string representation of the
UUID when we want to retrieve a given section. For consistency and
readability reasons, there should be only one representation.

In order to fix those issues, we now have a `SectionIdentifier`. This
identifier does not belong to the Db namespace and is passed everywhere
needed in the call stack and built as soon as possible when the user
query a section.

This `SectionIdentifier` has still a dependency on `Tuleap\Db\UUID`.
However its constructor is private, and its named constructor is
internal and can only be used by `SectionIdentifierFactory`. This should
limit misusage.

No functional changes.

Part of story #38288: add a new section

Change-Id: I6e154aa4565e9f3aad049629a0250fec22537e3d