refactor: Switch eslint to vue 3 and load styles through Vite assets
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
No functional change expected in cross-tracker widgets. Styles are still
loaded and the widgets function as before.

Why?
We had forgotten to add the widget to the eslint list of vue 3 apps, so
the vue 2 rules still applied. We can now remove superfluous <divs> that
only existed to have a unique root node in Vue 2 components. Note that
those rules require prop casing to be snake_case, so all prop names are
changed accordingly.
The styles can also be loaded by side-effect of the JS assets
(ViteAssets handle that). This will enable us to write the styles close
to the components in <style> tags.

Change-Id: Ifb65f4bd7de2349bcc6241bc5b76ca4072d83c93