feat: Insert new section after configuration
If you have an existing document that is not configured but already
contain sections, then when the add new section button is clicked, the
configuration modal is opened. However, once successfuly saved, the
modal can be closed but user has lost where the new section should be.

Now when the new selected tracker is compatible with the section
creation, then the form to enter the section is displayed as soon as the
configuration is saved.

If the selected tracker is not compatible, then the form is not
displayed.

Part of story #38288: add a new section

Change-Id: Iaa5d4301e13742a19a151fd739694297e956a6f1