fix: type issue in git/repositories-list
After these commits 96bf5771f3725c06f1face5e462ebfca5bc4b20c and
b7322be909b28762bf632081165c5e25d44bd582
We encountered a typing error in the `repository` props of
GitLabAdministration, which caused a typecheck error in composition API,
but was not a problem in class component.

The type of the props `repository` of GitLabAdministration has been
changed to FormattedGitLabRepository.

There was also a problem retrieving actions from the gitLab module
store, which is why we updated useAction to useNamespacedActions and
updated the test.

part of request #31168 Migration Vue2 => Vue3

Change-Id: Iccf2dee70e7c5c539106c9601744dca5668b2a7c