fix: prevent page leave with new section
To reproduce: create a new section, fill the form and save. Now click on
platform logo: you should not be asked for confirmation to leave the
page.

Having a boolean to count the number of open editors duplicates the
information: we already have a bunch of editors, we can ask if they are
open. That way there is less chance to miscalculate the number of open
editors.

Part of story #38288: add a new section

Change-Id: I89162c5fd6b2fcef970b1235cbea2d78f4af0e52