feat: Convert cog button to dropdown menu
With the arrival of PDF export feature, there will be more than one
action available on a document: PDF export & configuration. The cog
button is not anymore appropriate, we need to have a dropdown menu
instead.

If the user can edit the document, the dropdown is displayed and the
document can be configured.

If the user cannot edit the document, the dropdown is not displayed (for
now).

Part of story #38630: Export to file

Change-Id: I89ea667042b7f2656df474b4779efa371a15a948