fix: request #38668 All dates in the exported XLSX documents are moved to UTC TZ
We work around some changes made by the upstream to keep the dates
as they are displayed in web UI/retrieved in the REST API.

Instead of working with a `Date` that might get some transformation
during the work done by `xlsx` we work directly on the ISO strings
we get from the API. To prevent the dates being moved to UTC we cut
off the TZ info.

Introduced by 59c7d754b8ae69111a3617dbe4428897b78d27a3.

Change-Id: I16b77782723239b0bdbde3739d4a1f97f4f30299