feat: Introduce pdftemplate plugin
This plugin will allow site administrators (or those with sufficient
delegation permissions) to define PDF templates to be used later in
artidoc or other plugins wanting to export content in PDF format.

This plugin is flagged as experimental, but can be installed on dev
instances.

If the plugin is not installed, or not enabled, then nothing changes.

If the plugin is installed and enabled, then in artidoc there is an
option to export as PDF a document in the ellipsis menu. For now it only
calls `window.print()` to print the document, it will be enhanced later.

Part of story #38630: Export to file

Change-Id: Idaf66cd2e9a8ca2214ceba02ef53813ff4e53752