feat: apply styles defined in pdf template
The PDF template can provide its own style. Those style will be applied
when the user asks for PDF export.

We now use `print-js` library that will launch the print dialog on a
given dom element.

This dom element is responsible to give a printer version of the
document. This allow us to not have to worry about buttons, modals, or
other UI element that should not be printed.
It is hidden by default on the screen view.

Basic styles are added as an example. You can remove/change them to see
the change whenever you want to print the artidoc. Those styles will be
tailored later to have nice defaults.

Part of story #38630: Export to file

Change-Id: Id8cbb176e9918242e1dd4109c2ac45b799d781de