feat: Save configuration (table + default config)
Part of story #37941 Timetracking management widget

Saving the configration makes the widget query persistent when the page is reloaded.

This commit only creates a default configuration when the widget is added to the dashboard.
To do this, we created the main table for the query dates.
A Dao has been created which contains the creation and deletion methods required in Widget.

The primary key of the database query table is the widget id.
Other data are :
- the start date = as timestamp or null if a predefined time is set
- the end date = as timestamp or null if a predefined time is set
- the predefined period = as PredefinedTimePeriod type with 'last_7_days' as default value

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

How to test:
1. Add a timetracking management widget on your dashboard.
A new line should appear in your database with default values.

2. Remove the widget from your dashboard.
The line should be deleted from the database.

Change-Id: I8b52d2b75c801cce6ece81ef563515558a5706cd