feat: PDF template creation
Admin can now create a PDF template.

This template will be used if it is the first one. You will have to
hack the database to delete or reorganize your templates.

The new template comes with default css prefilled.

Preview button is not implemented yet but will allow to test if the
current settings match the expectations.

General styling code (css) is syntax highlighted thanks to CodeMirror.

Part of story #38630: Export to file

Change-Id: I3ce6d08d6446668af24c1744c580d2ca11bc0bea