feat: Display edit PDF template interface
This is an intermediate step before being able to edit a PDF template.
For now only the form is displayed.

To test: create a new template with a given label, description, and
style. Edit it => you should be able to change the label, description,
and style. If you click on Save button, you will reach a page not found
error since the update is not implemented yet. It will be part of a next
contribution.

Part of story #38630: Export to file

Change-Id: I28dc65af6f4a01cfeaa3c98deb1c856b4d459ed0