feat: Have a Docker image to serve tlp-doc-v2 content
This image is expected to be put behind a reverse proxy dealing with the
TLS terminaison.

The image is built with Nix, this way the image will at least be updated
when we bump the nixpkgs pin. Building from a Dockerfile and a `nginx`
base image is likely to not be kept updated on regular basis.

To test:
1. Build Storybook content with `TURBO_LOG_ORDER=grouped pnpm turbo run build-storybook --no-daemon --filter=@tuleap/tlp-doc-v2`
2. Build and start the Docker image with
   `docker load < $(nix-build --no-out-link ./src/scripts/tlp-doc-v2/container-image.nix) && docker run --rm -p 127.0.0.1:8080:8080 tlp-doc-v2`
3. Go to http://127.0.0.1:8080

Part of request #37211 Replace TLP-doc documentation

Change-Id: Ieb8537703b7772acc86c57ee4ac77e7bfd6862ba