feat: request #38682 Try to restart Tuleap worker services indefinitely
Instead of failing after a few try the workers will try to restart until
they can stay up. This means the system can heal itself after the DB was
not accessible for some time for example.

The RestartSec value has been increased from the default value (100ms by
default) to avoid a possible overload and because waiting a few more
seconds to restart the workers is not really an issue.

To test:
1. Copy the service file to /usr/lib/systemd/system/tuleap-worker@.service
2. Reload systemd with `systemctl daemon-reload`
3. Restart `tuleap` with `systemctl restart tuleap`
4. You can now simulate an issue by shuting down the DB and see the
   workers will try to reconnect

Change-Id: Id06533733909a1cfcd248fbcc55d05751c5e8100