feat: Preview PDF template
Administrator can now preview their changes in PDF templates.

The preview contain a fake document based on lorem ipsum text, a table,
and an image.

The image is big to see how the browser behaves and how we will be able
to manage later this type of images.

No functional changes in artidoc regarding the print of documents.

Part of story #38630: Export to file

Change-Id: Ib3a42d9d1ac2b2a17fecf9436cbc088d36bd2c75