feat: disable PDF option menu
When there is no template defined, then instead of removing the entry
menu, we disabled it and ask users to reach their administrator. That
way it does not hide the feature (else user don't see it exists and
won't search to print as PDF).

Furthermore, when the user is anonymous, then no template information is
send to the frontend and the entry is also disabled, so that users know
that they must log in. Anonymous users are not supposed to know every
bit of information of available templates since they can include
sensitive information.

If document is empty, or is in error, or sections are being loaded, or
sections are only pending ones, then the option is not displayed at all.

Took the opportunity to fix warning in console log by moving the
printer-version component.

Part of story #38630: Export to file

Change-Id: Ie2930818d2485fa999745823d6118d49472c9ac3