fix: Makes the resource active only if the feature flag is set
Part of story #37941 Timetracking management widget

The new resource TIMETRACKING_MANAGEMENT_WIDGET is enable for everyone.
It should only be enabled if the FEATURE_FLAG is set to 1.

This commit adds a check to ensure the FEATURE_FLAG is set to 1 before displaying the resource.

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

Change-Id: I97e067412cba6106090cfaa8818d7369573703da