feat: Build only docs for design system website
part of request #37211 Replace TLP-doc documentation

How to test:

1. Build Storybook content with `TURBO_LOG_ORDER=grouped pnpm turbo run
build-storybook --no-daemon --filter=@tuleap/tlp-doc-v2`
2. Build and start the Docker image with
`docker load < $(nix-build --no-out-link
./src/scripts/tlp-doc-v2/container-image.nix) && docker run --rm -p
127.0.0.1:8080:8080 tlp-doc-v2`
3. Go to http://127.0.0.1:8080
4. The sidebar features only the MDX documentation of the components,
the stories are not included anymore.

Why?
There is no demand for Stories themselves to be exposed on the public
documentation. It's actually the docs using the stories that are meant
to be published. We could add the stories later if the need arises.
Stories can still be accessed and used locally by developers, where they
are most useful (fast feedback loop).

Change-Id: Ia217d0a61d385a1c0973a3246e70690e067fd612