Fix request #38688 Tracker reports are not part of instrumentation process
The exit breaks the processing and, more importantly, the post
processing of the request.

This oneliner has a high potential of breakage in un-expected ways
because the exit is there since 2011 with the import of tracker v5 code
and I cannot say for sure what behavior was intended with this exit here
(there is no other exit in the controllers of that period of time so it
was there for a reason).

Change-Id: If863007289da915c03cbd7f3ff77bcb7b81e39fb