fix: print accurate content with new section
When a document is printed the PDF content should match saved sections
on the screen:
* Sections that are left untouched
* Sections that have been updated
* Sections that have been created
* Sections that have been created and updated

Issue was due to the fact that we created two separate editors for a
section: one for the screen, one for the print. The one for the print
was not aware of what was going on in the screen.
One way to fix that would have been to use a cached version of the
editor to share the same instance, but in fact print don't need at all
an editor. It only needs the title and description in read only.

Part of story #38630: Export to file

Change-Id: I6ad620fa0b78f1f95efde7e74a8c9867d6bdaf0d