fix: bump indentUnit to 4
The default css that is provided by Tuleap is indented with 4 spaces.
CodeMirror should have the same indentation width if users start to edit
the styles.

Go to your template, in an existing selector, add a new rule. It should be
indented like other rules.

Part of story #38630: Export to file

Change-Id: I83aab6b6517b09b8c480fcbed4fd50371302da17