fix: ensure uniqueness of id
In html, id attribute must be unique in a given DOM document.

Until now, we were using the id of the artifact. Using a number does not
guarantee that the id will be unique. Other element in the page might
have a similar id.

Decreasing the chance to have similar id is done by prefixing the id
with "section-" (we are targeting a section after all) and by using the
section uuid instead of the artifact. We don't know yet if there will be
other type of section than artifact (e.g.: freetext) so using the
artifact id is not really safe.

We can still navigate in the document via the table of contents. This
eventually breaks links that might have been bookmarked by early birds.

Not really part of story #38630: Export to file

Change-Id: Ic2c223a5b3318b5a60d110a625c155f5a363a0da