feat: add prose mirror editor to artidoc
You can choose between ckeditor and prose-mirror with the radio button if the feature flag is enabled.
In this patch, when you are in prose-mirror mode, you can click directly on section content or title and start editing without click to button "edit".
The button "cancel" and "save" will appear only when a change is occured, and disappear if you remove the added content.
You can save and a cancel a section like ckeditor.

Part of story : Implement prose-mirror editor

Change-Id: Ie53970ad78051f09a8a791cff11c209b41eb5840