feat: Simplify DOM structure
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
- Enable the feature flag:
  tuleap config-set feature_flag_enable_tql_select 1
- Write a query with a "SELECT", save the report.
- The numeric, text and date columns should be displayed. Text is
centered vertically in the cell.

Why?
I misunderstood the mockups and thought that the cells' content was
aligned to the right of each cell (like numeric tables). Since it turned
out that it's left-aligned, we do not need the div > span structure and
can drop some nodes. The DOM should be as simple as possible now.

Change-Id: I8d9fa8029cf4b564cf7a75f0cb846055aa1866cb