feat: Save widget configuration (users)
Part of story #37941 Timetracking management widget

Update the route REST to update query with user ids.

The widget needs to save a list of users, so we create a new table named `plugin_timetracking_management_query_users`.
This table contains two columns, one for the widget id and one for the user id.
The idea is to update the list of users at the same time as we save the query with dates or predefined time periods.
A check is done before saving. If one of the ids int the list is unknown, it returns a Fault.

How to test:
-- Add a timetracking management widget to you dashboard.
-> Your `plugin_timetracking_management_query_users` table should be empty.
-- In the api explorer, update the query by adding users (you can use the example).
-> Your database should then contained the corresponding user ids.
-- Update the query again, changing user ids in the api explorer.
-> Your database should be updated.

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

Change-Id: I4bd5bd96de2a1a324aba76c8e0603fb28e4aa182