feat: Check permissions
Part of story #37941 Timetracking management widget

Ensure that the widget id exists.
Ensure that the current user is the widget owner.

If the widget id does not exist, or if the current user is not the widget owner, a 'Widget not found' fault is returned.

How to test:
-- Add a timetracking management widget to your dashbord.
-- With a diffrent user, try to update the query through the API explorer.
-> It should display a 404 error with the message "Not Found: Widget not found"
-- Try to update an unknown widget (e.g widget_id = -66).
-> It should display a 404 error with the message "Not Found: Widget not found"

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

Change-Id: I0ad78acb8e2b8cdbfe6350dcf08a9ebd28981325